requirements.txt
aicommit==0.1.0
aiohttp==3.8.5
aiosignal==1.3.1
ansicon==1.89.0
asgiref==3.5.1
async-timeout==4.0.3
attrs==23.1.0
black==22.12.0
blessed==1.19.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.4
cryptography==39.0.2
dataclasses-json==0.5.14
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.53.0
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.2.0
django-jalali-date==1.0.1
django-restframework==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
environs==9.5.0
et-xmlfile==1.1.0
fire==0.4.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
gunicorn==20.1.0
halo==0.0.31
idna==3.4
importlib-metadata==5.2.0
jaraco.classes==3.2.3
jdatetime==4.1.1
jinxed==1.2.0
keyring==23.13.1
langchain==0.0.272
langsmith==0.0.26
log-symbols==0.0.14
Markdown==3.4.4
marshmallow==3.19.0
more-itertools==9.0.0
multidict==6.0.4
mypy-extensions==0.4.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numexpr==2.8.5
numpy==1.24.3
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
packaging==23.0
pandas==1.5.3
pandas-stubs==1.5.2.230105
pathspec==0.10.3
Pillow==9.1.1
platformdirs==2.6.2
prompt-toolkit==1.0.14
protobuf==3.20.3
psycopg2-binary==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pygit2==1.11.1
Pygments==2.13.0
PyInquirer==1.0.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-editor==1.0.4
python3-openid==3.2.0
pytz==2022.7.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
readchar==2.0.1
redis==4.4.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
smmap==5.0.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
spinners==0.0.24
SQLAlchemy==1.4.49
sqlparse==0.4.2
tenacity==8.2.3
termcolor==1.1.0
tomli==2.0.1
tqdm==4.64.1
types-pytz==2022.7.0.0
typing-inspect==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==3.0
urllib3==1.26.16
wcwidth==0.2.5
websockets==10.0
yarl==1.9.2
GDAL==3.4.3
# GDAL @ file:///C:/Users/novin/Desktop/GDAL-3.4.3-cp310-cp310-win_amd64.whl#sha256=9893da5532a6c5dadc1fc3bc402ae29d161b7cf190d778f129da84d0e50baa29


